requirements.txt
streamlit
pandas
numpy
google-generativeai
python-dotenv
langchain-google-genai
langchain
